% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGData.R
\name{createGData}
\alias{createGData}
\title{Create a gData object}
\usage{
createGData(geno, map, kin, pheno, covar)
}
\arguments{
\item{geno}{matrix or data.frame with genotypes in the rows and markers in the columns.
If no row names are used they are taken from \code{pheno} (if supplied and dimension matches).
If no column names are used the row names are taken from \code{map}
(if supplied and dimension matches).}

\item{map}{a data.frame with columns chr for chromosome and pos for position. Positions can
be in basepair or centimorgan. They should not be cumulative over the chromosomes.
Other columns are ignored. Marker names should be in the row names. These should match
the marker names in \code{geno} (if supplied).}

\item{kin}{a kinship matrix with genotype in rows and colums. These should be identical to the genotypes
in \code{geno}}

\item{pheno}{a matrix or data.frame or a list of matrices/data.frames with phenotypic data, with
genotypes in the rows and traits in the columns. A list of data.frames can be used for replications,
i.e. different environments.}

\item{covar}{a data.frame with extra covariates per genotype. Genotype should be in the rows.}
}
\value{
an object of class gData with the following components:
\itemize{
\item{map a data.frame containing map data. Map is sorted by chromosome and position.}
\item{markers a matrix containing marker information}
\item{pheno a list of matrices containing phenotypic data}
\item{kinship a kinship matrix}
\item{covar a data.frame with extra covariates}
}
}
\description{
\code{createGData} creates an object of S3 class with genotypic and phenotypic data for usage in
further analysis. All input to the function is optional, however at least one input should be
provided.
}
