% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estEffs.R
\name{estEffs}
\alias{estEffs}
\title{Estimates for covariates}
\usage{
estEffs(Y, W = matrix(data = 1, nrow = nrow(Y)), X, Vg, Ve, K,
  returnSe = TRUE, estCom = FALSE, nChunks = min(ncol(X),
  ceiling(ncol(Y) * ncol(W) * ncol(X)/50000)))
}
\arguments{
\item{Y}{An n x p matrix of observed phenotypes, on p traits or environments
for n genotypes. No missing values are allowed.}

\item{W}{An n x c covariate matrix, c being the number of covariates and n
being the number of genotypes. c has to be at least one (typically an
intercept). No missing values are allowed.}

\item{X}{An n x ns matrix of marker scores. Neither missing values nor
non-segregating markers are allowed.}

\item{Vg}{A p x p matrix of genetic covariances.}

\item{Ve}{A p x p matrix of environmental covariances.}

\item{K}{An n x n genetic relatedness matrix.}

\item{returnSe}{Should standard errors and p-values be returned?}

\item{estCom}{Should the common SNP-effect model be fitted?}

\item{nChunks}{An integer, the number of blocks in which the calculations
should be split.}
}
\value{
A list containing the estimates, optionally the standard errors of
the estimates and corresponding p-values. If \code{estCom = TRUE} also
common SNP-effects, their standard errors and corresponding p-values and
the p-values for QtlxE are output.
}
\description{
Compute the estimates and standard errors for the covariates in the input
matrix W.
}
\references{
Zhou, X. and Stephens, M. (2014). Efficient multivariate linear
mixed model algorithms for genome-wide association studies. Nature Methods,
February 2014, Vol. 11, p. 407–409
}
\keyword{internal}
