% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estEffTot.R
\name{estEffTot}
\alias{estEffTot}
\title{Helper function for estimating effect}
\usage{
estEffTot(
  markers,
  map,
  minCofactorProximity = 50,
  X,
  Y,
  K,
  XRed,
  Vg,
  Ve,
  VgRed,
  VeRed,
  snpCov,
  allFreq,
  MAF,
  estCom,
  nCores = NULL
)
}
\description{
Function is a wrapper around estEffsCPP, useful for usage in chromosome
specific calculations.
}
\keyword{internal}
