% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastGLS.R
\name{fastGLS}
\alias{fastGLS}
\alias{fastGLSCov}
\title{fastGLS}
\usage{
fastGLS(Y, X, Sigma)

fastGLSCov(y, X, Sigma, covs, nChunks = 10)
}
\arguments{
\item{X}{an n x m matrix of marker-scores, n being the number of individuals, m the number of markers.
no missing values allowed.}

\item{Sigma}{an n x n covariance matrix. No missing values allowed.}

\item{y}{a numeric vector of length n of phenotypic scores. No missing values allowed.}

\item{covs}{an n x c matrix of covariates (NOT including an intercept). No missing values allowed.}

\item{nChunks}{an integer, the number of parts in which the calculations are split.}
}
\value{
a list containing the following components:
\itemize{
\item{\code{pValues} a vector of p-values for the GLS F-test}
\item{\code{beta} a vector of effect sizes}
\item{\code{betaSe} a vector of standard errors of the effect sizes}
\item{\code{RLR2} a vector of R_LR^2 statistics as defined in Sun et al.}
}
}
\description{
compute p-values for the GLS F-test as in emma-x, Fast-LMM or scan_GLS. \code{fastGLS} can be used when
there are no further covariates, otherwise use \code{fastGLSCov}.
}
\references{
Segura et al. (2012) An efficient multi-locus mixed-model approach for genome-wide association
studies in structured populations. Nature Genetics, June 2012, Vol. 44, p. 825–830

Sun et al. (2010) Variation explained in mixed-model association mapping.
Heredity, February 2010, Vol. 105, p. 333–340
}
