% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastGLSIBD.R
\name{fastGLSIBD}
\alias{fastGLSIBD}
\title{Fast Generalized Least Squares algoritm for IBD based QTL mapping}
\usage{
fastGLSIBD(y, X, Sigma, covs, ref)
}
\arguments{
\item{y}{A numeric vector of length n of phenotypic scores. No missing
values allowed.}

\item{X}{An array of marker probabilities.}

\item{Sigma}{An n x n covariance matrix. No missing values allowed.}

\item{covs}{An n x c matrix of covariates (NOT including an intercept).
No missing values allowed.}

\item{ref}{An integer indicating the allele to use as reference allele in
the computations.}
}
\description{
Compute statistics for the Generalized Least Squares (GLS) F-test for
IBD based QTL Mapping.
}
\keyword{internal}
