% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gData.R
\name{gData}
\alias{gData}
\title{Create a gData object}
\usage{
gData(geno, genoHeader = TRUE, genoRowNames = TRUE, map, mapSnpName = 1,
  mapChromosome = 2, mapPosition = 3, kin = NULL, pheno = NULL,
  phenoGeno = 1, createExternal = "none")
}
\arguments{
\item{geno}{string, specifying a csv file with the genotypic data, markers in the rows and
genotypes in the columns. Alternatively, an dataframe with a similar layout.}

\item{genoHeader}{does \code{geno} contain a header row. If \code{FALSE} genotypes are given a default
name.}

\item{genoRowNames}{does \code{geno} contain rownames in its first column. If \code{FALSE} markers
are given a default name.}

\item{map}{string, specifying a csv file with at least columns for SNP name, chromosome and position.
The positions should be in base-pair or morgan. They should not be cumulative over the chromosomes. Other
columns are ignored. Alternatively, an dataframe with a similar layout.}

\item{mapSnpName}{the column corresponding to the SNP name in \code{map}, either a string or the
column number. If \code{NULL} the first column is assumed to be the SNP name.}

\item{mapChromosome}{the column corresponding to the cromosome number in \code{map}, either a string or the
column number. If \code{NULL} the second column is assumed to be the chromosome number.}

\item{kin}{string, specifying a csv file with the kinship matrix, with genotypes in both the rows
and the columns. Alternatively, an dataframe with a similar layout. Row names and column names should
be identical to column names in geno. If \code{NULL} a kinship matrix is computed from \code{geno}
using GRM.}

\item{pheno}{string, specifying a csv file with phenotypic data, with genotypes in both the rows
and the columns. Alternatively, an dataframe with a similar layout. Row names and column names should
be identical to column names in geno.}

\item{phenoGeno}{the column corresponding to the genotype in \code{pheno}, either a string or the
column number. If \code{NULL} the first column is assumed to be the genotype.}

\item{position}{the column corresponding to the position in \code{map}, either a string or the
column number. If \code{NULL} the third column is assumed to be the position.}
}
\value{
an object of class gData with the following components:
\itemize{
\item{map a dataframe containing map data}
\item{markers a dataframe containing marker information}
\item{pheno a dataframe containing phenotypic data}
\item{kinship a kinship matrix}
\item{genotypes a charcter vector of genotypes}
\item{chromosomes a numeric vector of chromosome numbers}
\item{nChr the number of chromosomes}
\item{nGeno the number of genotypes}
\item{nSNP the number of SNPs}
\item{chrLengthsBp a numeric vector of chromosome lengths}
}
}
\description{
Create an object of S3 class gData based on external files or dataframes containing genotypic and phenotypic
information
}
