% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gDataDropsRestr}
\alias{gDataDropsRestr}
\title{Subset of DROPS data for use in examples}
\format{
An object of class \code{gData} of length 5.
}
\source{
\doi{10.15454/IASSTN}
}
\usage{
gDataDropsRestr
}
\description{
A \code{\link[statgenGWAS]{gData}} object based on a subset of the DROPS data
set used in the statgenGWAS package. The data is restricted to 3 traits and
10\% (just over 4000) of the available markers. For a full description of the
data set see \code{\link[statgenGWAS]{dropsData}}.
}
\references{
Millet, E. J., Pommier, C., et al. (2019). A multi-site
experiment in a network of European fields for assessing the maize yield
response to environmental scenarios - Data set.
\doi{10.15454/IASSTN}

Ganal MW, et al. (2011) A Large Maize (Zea mays L.) SNP
Genotyping Array: Development and Germplasm Genotyping, and Genetic Mapping
to Compare with the B73 Reference Genome. PLoS ONE 6(12): e28334.
\doi{10.1371/journal.pone.0028334}
}
\keyword{datasets}
