% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicPrediction.R
\name{genomicPrediction}
\alias{genomicPrediction}
\title{Perform genomic prediction}
\usage{
genomicPrediction(gData, traits = NULL, environments = NULL, training,
  kin = NULL, kinshipMethod = c("astle", "GRM", "IBS", "vanRaden"),
  keep = NULL)
}
\arguments{
\item{gData}{An object of class \code{gData} containing at least \code{map},
\code{markers} and \code{pheno}.}

\item{traits}{A vector of traits on which to run GWAS. These can be either
numeric indices or character names of columns in \code{pheno}. If \code{NULL}
GWAS is run on all traits.}

\item{environments}{A vector of environments on which to run GWAS. These can
be either numeric indices or character names of list items in \code{pheno}.
If \code{NULL} GWAS is run for all environments.}

\item{training}{A character vector with names of genotypes to use as
training data for the prediction.}

\item{kin}{An optional kinship matrix or list of kinship matrices. These
matrices can be from the \code{matrix} class as defined in the base package
or from the \code{dsyMatrix} class, the class of symmetric matrices in the
Matrix package.\cr
If \code{GLSMethod} = "single" then one matrix should be provided, if
\code{GLSMethod} = "multi" a list of chromosome specific matrices of lenght
equal to the number of chromosomes in \code{map} in \code{gData}.\cr
If \code{NULL} then matrix \code{kinship} in \code{gData} is used. \cr
If both \code{kin} is provided and \code{gData} contains a matrix
\code{kinship} then \code{kin} is used.}

\item{kinshipMethod}{An optional character indicating the method used for
calculating the kinship matrix(ces). Currently "astle" (Astle and Balding,
2009), "GRM", "IBS" and "vanRaden" (VanRaden, 2008) are supported. If a
kinship matrix is supplied either in \code{gData} or in parameter \code{kin}
\code{kinshipMethod} is ignored.}

\item{keep}{A character vector with names of columns in \code{covar} in
\code{gData} that should be included in the output.}
}
\value{
A data.frame containing predictions for genotypes per trait.
Columns in \code{keep} are included in the output.
}
\description{
Perform genomic prediction
}
\references{
Piepho, H.-P., Möhring, J., Schulz-Streeck, T. and Ogutu, J. O.
(2012), A stage-wise approach for the analysis of multi-environment trials.
Biom. J., 54: 844–860. doi:10.1002/bimj.201100219
}
