% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGWAS.R
\name{GWAS}
\alias{GWAS}
\alias{createGWAS}
\alias{is.GWAS}
\title{S3 Class GWAS}
\usage{
createGWAS(GWAResult = NULL, signSnp = NULL, kin = NULL,
  thr = NULL, GWASInfo = NULL)

is.GWAS(x)
}
\arguments{
\item{GWAResult}{An optional data.frame or list of data.frames containing the
overall analysis results. Should a least contain columns \code{trait}, the
evaluated trait, \code{snp}, the name of the SNP, \code{chr}, the chromosome
number, \code{pos}, the position of the SNP on the chromosome,
\code{pValue}, the p-values from the analysis and \code{LOD} the LOD-score.}

\item{signSnp}{An optional data.frame or list of data.frames containing
information on the significant SNPs and optionally the SNPs close to the
significant SNPs. Should at least contain columns \code{trait}, the
evaluated trait, \code{snp}, the name of the SNP, \code{pValue}, the p-values
from the analysis and \code{LOD} the LOD-score.}

\item{kin}{An optional kinship matrix or list of chromosome specific kinship
matrices.}

\item{thr}{An optional numerical value, the threshold used in performing the
GWAS analysis.}

\item{GWASInfo}{a list containing extra information concering the GWAS
analysis.}

\item{x}{An \code{R} object}
}
\value{
\code{createGWAS} returns an object of class GWAS, a list of the
input items.\cr\cr
\code{is.GWAS} returns \code{TRUE} or \code{FALSE} depending on whether its
argument is a \code{GWAS} object.
}
\description{
\code{createGWAS} creates an object of S3 class GWAS containing the results
of a GWAS analysis.
\code{GWAResult} and \code{signSnp} are both optional, however at least one
of those should be provided as input.\cr
\code{summary} and \code{plot} functions are available.\cr\cr
\code{is.gData} tests if an \code{R} object is of class \code{gData}.
}
\seealso{
\code{\link{summary.GWAS}}, \code{\link{plot.GWAS}}
}
