% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kinshipTransform.R
\name{kinshipTransform}
\alias{kinshipTransform}
\title{kinshipTransform}
\usage{
kinshipTransform(K)
}
\arguments{
\item{K}{an n by n matrix}
}
\value{
the ...
}
\description{
Helper function for computing ... Used for calculating kinship matrices.
}
