% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LLDiag.R
\name{LLDiag}
\alias{LLDiag}
\title{Compute log-likelihood}
\usage{
LLDiag(Y, X = data.frame(), VArray, VInvArray)
}
\arguments{
\item{Y}{a p x n matrix of observed phenotypes, on p traits or environments for n genotypes.
See details.}

\item{X}{an optional c x n covariate matrix, c being the number of covariates and n being the number
of genotypes. c has to be at least one (typically an intercept). No missing values are allowed.}

\item{VArray}{an n x p x p dimensional array obtained as an output from the function
\code{\link{makeVArray}}. It contains for each genotype l the p x p matrix \eqn{v_l} (in
the notation of Zhou and Stephens)}

\item{VInvArray}{an n x p x p dimensional array obtained as an output from the function
\code{\link{makeVInvArray}}. It contains for each genotype l the p x p matrix \eqn{v_l ^ {-1}} (in
the notation of Zhou and Stephens)}
}
\value{
a numeric value for the \eqn{t(y) * P * y} part of the log-likelihood function.
}
\description{
Compute \eqn{t(y) * P * y}, part of the log-likelihood functions from equation 26 and 27
in Zhou and Stephens (2014) using equation 50. Equation 56, 57 and 58 are used to do
the actual computations.
}
\details{
It is assumed that X and Y have already been rotated by Uk, where Uk is such that
the kinship matrix K equals \eqn{K = Uk * Dk * t(Uk)}.\cr
The original X and Y are right multiplied by Uk, e.g. \code{Y <- Y * Uk}. See Zhou
and Stephens (2014), supplement.\cr
It is these rotated versions that are the input of this function.
}
\references{
Zhou, X. and Stephens, M. (2014). Efficient multivariate linear mixed model algorithms for
genome-wide association studies. Nature Methods, February 2014, Vol. 11, p. 407–409
}
\keyword{internal}
