% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LLQuadFormDiagCPP}
\alias{LLQuadFormDiagCPP}
\title{Compute tYPY as in Zhou and Stephens eqn. 50.}
\usage{
LLQuadFormDiagCPP(y, vInv, size_param = NULL)
}
\value{
A numerical value for the \eqn{t(y) * P * y} part of the
log-likelihood function.
}
\description{
Compute \eqn{t(y) * P * y}, part of the log-likelihood functions from
equation 26 and 27 in Zhou and Stephens using equation 50. Equation 56, 57
and 58 are used to do the actual computations.
}
\details{
It is assumed that X and Y have already been rotated by Uk, where Uk is such
that the kinship matrix K equals \eqn{K = Uk * Dk * t(Uk)}.\cr
The original X and Y are right multiplied by Uk, e.g. \code{Y <- Y * Uk}.
See Zhou and Stephens (2014), supplement.\cr
It is these rotated versions that are the input of this function.
}
\references{
Zhou, X. and Stephens, M. (2014). Efficient multivariate linear
mixed model algorithms for genome-wide association studies. Nature Methods,
February 2014, Vol. 11, p. 407–409
}
\keyword{internal}
