% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRTTest.R
\name{LRTTest}
\alias{LRTTest}
\title{Test for marker effect.}
\usage{
LRTTest(X, x, Y, VInvArray, SS0 = NULL)
}
\arguments{
\item{X}{a c x n covariate matrix, c being the number of covariates and n being the number
of genotypes. c has to be at least one (typically an intercept). See details.}

\item{x}{an additional covariate, usually a single marker, a numeric vector of length n. See details.}

\item{Y}{a p x n matrix of observed phenotypes, on p traits or environments for n genotypes.
See details.}

\item{VInvArray}{an n x p x p dimensional array obtained as an output from the function
\code{\link{makeVInvArray}}. It contains for each genotype l the p x p matrix \eqn{v_l ^ {-1}} (in
the notation of Zhou and Stephens)}

\item{SS0}{the log-likelihood value of a NULL model. This is the model with trait specific means
only and without an additional marker effect. If \code{NULL} then this value is calculated.}
}
\value{
A list containing the test statistics \code{pvalue}, \code{FStat} and \code{df}. Furthermore
the Log-likelihood of the null model and the alternative model, \code{SS0} and \code{SS1}. Also the
estimated effect, its standard error and results of the wald test for the additional covariate a returned.
}
\description{
Test the effect of an extra covariate. This is done by testing the results of a NULL model
only containing the trait specific means and an alternative model containing the trait
specific marker effect as well.
}
\details{
No missing values are allowed in X, Y and x.\cr
It is assumed that X, Y and x have already been rotated by Uk, where Uk is such that
the kinship matrix K equals \eqn{K = Uk * Dk * t(Uk)}.\cr
The original X, Y and x are right multiplied by Uk, e.g. \code{Y <- Y * Uk}. See Zhou
and Stephens 2014, supplement.\cr
It is these rotated versions that are the input of this function.
}
\references{
Zhou, X. and Stephens, M. (2014). Efficient multivariate linear mixed model algorithms for
genome-wide association studies.
}
