% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVArray.R
\name{makeVArray}
\alias{makeVArray}
\alias{makeVInvArray}
\title{Create array of (inverted) variance matrices}
\usage{
makeVArray(Vg, Ve, Dk)

makeVInvArray(Vg, Ve, Dk)
}
\arguments{
\item{Vg}{a p x p symmetric matrix of genetic variance components}

\item{Ve}{a p x p symmetric matrix of environmental variance components}

\item{Dk}{a vector of length n containing the eigenvalues obtained by the
eigen-decomposition of the kinship matrix K: \eqn{K = Uk * Dk * t(Uk)}}
}
\value{
A list of n p x p matrices \eqn{v_l} where \eqn{v_l =
Dk_{l,l} * Vg + Ve \forall l = 1, ..., n}.\cr
When using \code{makeVInvArray} the output matrices are inverted.
}
\description{
As in Zhou and Stephens, supplement page 13, create the array of (inverted)
variance matrices for the transformed genotypes per individual as given by
equation 4.
}
\references{
Zhou, X. and Stephens, M. (2014). Efficient multivariate linear
mixed model algorithms for genome-wide association studies. Nature Methods,
February 2014, Vol. 11, p. 407–409.
}
\keyword{internal}
