% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMFA.R
\name{makeVLst}
\alias{makeVLst}
\alias{makeVInvLst}
\title{Create a list of (inverted) variance matrices}
\usage{
makeVLst(Vg, Ve, Dk)

makeVInvLst(Vg, Ve, Dk)
}
\arguments{
\item{Vg}{A p x p symmetric matrix of genetic variance components}

\item{Ve}{A p x p symmetric matrix of environmental variance components}

\item{Dk}{A vector of length n containing the eigenvalues obtained by the
eigen-decomposition of the kinship matrix K: \eqn{K = Uk * Dk * t(Uk)}}
}
\value{
A list of n p x p matrices \eqn{v_l} where \eqn{v_l =
Dk_l * Vg + Ve \forall l = 1, ..., n}.\cr
When using \code{makeVInvLst} the output matrices are inverted.
}
\description{
As in Zhou and Stephens, supplement page 13, create the list of (inverted)
variance matrices for the transformed genotypes per individual as given by
equation 4.
}
\references{
Zhou, X. and Stephens, M. (2014). Efficient multivariate linear
mixed model algorithms for genome-wide association studies. Nature Methods,
February 2014, Vol. 11, p. 407–409.
}
\keyword{internal}
