% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixPlot.R
\name{matrixPlot}
\alias{matrixPlot}
\title{Create a matrix plot}
\usage{
matrixPlot(
  effectDat,
  signSnp,
  map,
  xLab = "Chromosomes",
  yLab = "Parent",
  chrBoundaries,
  founders,
  ...,
  output = TRUE
)
}
\description{
Create a matrix plot of effect sizes and directions per parent for
significant snps.
}
\keyword{internal}
