% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixRoot.R
\name{matrixRoot}
\alias{matrixRoot}
\title{Compute Square root of a symmetric, positive definite matrix}
\usage{
matrixRoot(X)
}
\arguments{
\item{X}{a symmetric, positive definite matrix.}
}
\value{
a matrix Y such that \eqn{Y^2 = X}.
}
\description{
Given an symmetric, positive definite matrix X a matrix Y is computed such
that \eqn{Y^2 = X}. Computation is done using eigendecomposition of X.
}
\keyword{internal}
