% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.GWAS.R
\name{plot.GWAS}
\alias{plot.GWAS}
\title{Plot function for the class \code{GWAS}}
\usage{
\method{plot}{GWAS}(x, ..., type = c("manhattan", "qq", "qtl"),
  environment = NULL, trait = NULL)
}
\arguments{
\item{x}{An object of class \code{GWAS}.}

\item{...}{further arguments to be passed on to the actual plotting
functions.}

\item{type}{A character string indicating the type of plot to be made. One
of "manhattan", "qq" and "qtl".}

\item{environment}{A character string or numeric index indicating for which
environment the plot should be made. If \code{x} only contains results for
one trait \code{environment} may be \code{NULL}.}

\item{trait}{A character string indicating for which trait the results
should be plotted. For \code{type} "qtl" all traits are plotted. If \code{x}
only contains results for one trait \code{trait} may be \code{NULL}.}
}
\description{
Creates a plot of an object of S3 class \code{GWAS}. Three types of plots
can be made:
\itemize{
\item{a manhattan plot using the function \code{\link{manhattanPlot}}}
\item{a qq plot using the function \code{\link{qqPlot}}}
\item{a qtl plot using the function \code{\link{qtlPlot}}}
}
}
\details{
When making a manhattan plot all markers in the \code{GWAResult} data.frame
in the \code{GWAS} object are plotted. Significant SNPs are taken from the
\code{signSnp} data.frame in the \code{GWAS} object and marked in the plot.
Also the LOD-threshold is taken from the \code{GWAS} object and plotted as
a horizontal line. Both \code{signSnp} and \code{thr} can be left empty and
will then be ignored in the plot.\cr
\code{...} can be used to pass extra arguments to the actual plotting
functions. See those fuctions for details.
}
\seealso{
\code{\link{manhattanPlot}}, \code{\link{qqPlot}},
\code{\link{qtlPlot}}
}
