% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqPlot.R
\name{qqPlot}
\alias{qqPlot}
\title{qq-plot of observed versus expected LOD-scores}
\usage{
qqPlot(pValues, ..., output = TRUE)
}
\arguments{
\item{pValues}{A numeric vector of pValues. Missings are ignored when
plotting.}

\item{...}{Other graphical parameters passed on to the actual plot function.}

\item{output}{Should the plot be output to the current device? If
\code{FALSE} only a list of ggplot objects is invisibly returned.}
}
\description{
Given a vector of pvalues, generate a qq-plot of observed LOD-scores versus
expected LOD-scores. Code taken from Segura et al. and adapted
}
\references{
Segura et al. (2012) An efficient multi-locus mixed model
approach for genome-wide association studies in structured populations
}
