% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtlPlot.R
\name{qtlPlot}
\alias{qtlPlot}
\title{Draw a qtl plot}
\usage{
qtlPlot(data, chromosome = "chromosome", trait = "trait",
  snpEffect = "effect", snpPosition = "position", map, normalize = FALSE,
  sortData = FALSE, yLab = "Traits", exportPptx = FALSE, pptxName = "")
}
\arguments{
\item{data}{QTL data to be plotted}

\item{chromosome}{a character indicating the data column containing the chromosome number.}

\item{trait}{a character indicating the data column containing the trait name.}

\item{snpEffect}{a character indicating the data column containing the snp effect.}

\item{snpPosition}{a character indicating the data column containing the position of the snp on
the chromosome.}

\item{map}{a dataframe with at least the columns chromosome, the number of the chromosome
and position, the position of the snp on the chromosome. These are used for calculating the
fysical limits of the chromosome.}

\item{normalize}{should the snpEffect be normalized?}

\item{sortData}{should the data be sorted before plotting? Either false if no sorting should be done
or a character indicating the data column on which the data should be sorted.}

\item{yLab}{y-axis label.}

\item{exportPptx}{should the plot be exported to a .pptx file.}

\item{pptxName}{name of the .pptx file to which the plot is exported. Ignored if exportPptx =
\code{FALSE}}
}
\value{
a plot.
}
\description{
Draw a qtl plot
}
\references{
Millet et al. (2016) Genome-wide analysis of yield in Europe: Allelic effects vary
with drought and heat scenarios
}
