% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readIBDProbs}
\alias{readIBDProbs}
\title{Read IBD probabilities}
\usage{
readIBDProbs(infile)
}
\arguments{
\item{infile}{A character string, a link to a .csv file with IBD
probabilities.}
}
\value{
A list with two components:
\itemize{
\item{markArr} {A three-dimensional array containing the IBD probabilities
with genotypes in the rows, markers in the columns and founders in the
third dimension.}
\item{map} {A data.frame with two columns, chr(omosome) and pos(ition) with
genotypes as rownames.}
}
Both markArr and map can be used directly in \code{\link{createGData}}.
}
\description{
Read a file with IBD probabilities computed by the RABBIT software package.
}
