% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runEmma.R
\name{runEmma}
\alias{runEmma}
\title{Compute REML estimates of variance components using EMMA algorithm.}
\usage{
runEmma(gData, trait, environment, K = NULL, covar = NULL, snpName = NULL,
  Z = NULL, nGrids = 100, lLim = -10, uLim = 10, eps = 1e-10)
}
\arguments{
\item{gData}{an object of class gData containing at least a data.frame \code{pheno}. If \code{K} is
not supplied a matrix \code{kinship} should be in \code{gData}. If covariates are included
then a data.frame covar is needed as wel and if an extra snp is to be included as covariate
(defined in \code{snpName}) then a data.frame \code{markers} is also needed. Missing values
in \code{pheno} are allowed but will be excluded from the calculations.}

\item{trait}{a trait for which to estimate variance components. This can be either numeric index
or character name of a column in \code{pheno}.}

\item{environment}{an environment for which to estimate variance components. This can be either numeric
index or character name of a list item in \code{pheno}.}

\item{K}{an optional kinship matrix. If \code{NULL} then matrix \code{kinship} in \code{gData} is used.
If both \code{K} is provided and \code{gData} contains a matrix \code{kinship} then \code{K} is used.}

\item{covar}{an optional vector of covariates taken into account when estimating variance components.
These can be either numeric indices or character names of columns in \code{covar} in \code{gData}.
If \code{NULL} no covariates are used.}

\item{snpName}{an optional character name of a marker in \code{markers} in \code{gData} to be
included as covariate. If used the \code{gData} object should contain a data.frame \code{markers}.}

\item{Z}{an optional incidence matrix mapping each observed phenotype to one of inbred strains.}

\item{nGrids}{an integer indicating the number of intervals used for local optimisation within the
algorithm.}

\item{lLim}{a numeric value indicating the lower limit of the interval over which optimisating is
done.}

\item{uLim}{a numeric value indicating the upper limit of the interval over which optimisating is
done.}

\item{eps}{a numeric value used as computational tolerance in the algorithm.}
}
\value{
a list with two components:
\itemize{
\item{\code{varcomp} a vector of genetic variance Vg and residual variance Ve}
\item{\code{K} the kinship matrix used in the algorithm (entries for missing values in the
original kinship matrix are filtered out.)}
}
}
\description{
Using the EMMA algorithm as is Kang et al. (2008) compute REML estimates of genetic and residual
variance components.
}
\references{
Kang et al. (2008) (Efficient Control of Population Structure in Model Organism
Association Mapping. Genetics, March 2008, Vol. 178, no. 3, p. 1709-1723
}
\keyword{internal}
