% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGWAS.R
\name{runGWAS}
\alias{runGWAS}
\title{run GWAS}
\usage{
runGWAS(markers, map, Y, K, X = NULL, subsetMarkers = FALSE,
  markerSubset = "", snpCovariates = "", MAF = 0.05, fitVarComp = TRUE,
  covModel = 1, vEDiag = TRUE, tolerance = 1e-06, maxIter = 2e+05,
  maxDiag = 10000, mG = 1, mE = 1, CmHet = TRUE, DmHet = TRUE,
  stopIfDecreasing = TRUE, computeLogLik = TRUE, Vg = NULL, Ve = NULL,
  reduceK = FALSE, nPca = NULL)
}
\arguments{
\item{markers}{A dataframe with marker information. Row names should be chromosomes, column names
genotypes.}

\item{map}{A dataframe with four columns; chromosome, the number of the chromosome,
position, the position of the snp on the chromosome, snp.name, the name of the snp and
cum.position, the cumulative position of the snp starting from the first chromosome.}

\item{Y}{an n x p matrix or dataframe of observed phenotypes, on p traits or environments for n
individuals. Missing values are allowed.}

\item{K}{an n x n kinship matrix.}

\item{X}{an n x c covariate matrix, c being the number of covariates and n being the number
of genotypes. c has to be at least one (typically an intercept). No missing values are allowed.
If not provided a vector of 1s is used.}

\item{subsetMarkers}{should the marker data be subsetted?}

\item{markerSubset}{numeric or character vector for subsetting the markers. Ignored if
subsetMarkers = \code{FALSE}.}

\item{snpCovariates}{a character vacter of snps that are to be included as covariates.}

\item{MAF}{a numeric value between 0 and 1. Snps with a minor allele frequency outside MAF
and 1 - MAF are excluded from the GWAS analysis.}

\item{fitVarComp}{should the variance components be fitted? If \code{FALSE} they should be supplied
in Vg and Ve}

\item{covModel}{an integer value for the model used when fitting the variance components.
\enumerate{
\item{unstructured for both Vg and Ve (as in Zhou and Stephens)}
\item{unstructered for both Vg and Ve (pairwise, as in Furlotte and Eskin)}
\item{factor-analytic for both Vg and Ve}
\item{approximate ...}
}
Ignored if fitVarComp = \code{FALSE}}

\item{vEDiag}{Should there be environmental correlations if covModel = 2? If traits are measured on
the same individuals put \code{FALSE}.}

\item{tolerance}{a numerical value. Used when fitting the factor analytical model if covModel = 3.
See \code{\link{EMFA}}.}

\item{maxIter}{an integer. Used when fitting the factor analytical model if covModel = 3.
See \code{\link{EMFA}}.}

\item{maxDiag}{a numerical value. Used when fitting the factor analytical model if covModel = 3.
See \code{\link{EMFA}}.}

\item{mG}{an integer. Used when fitting the factor analytical model if covModel = 3.
See \code{\link{EMFA}}.}

\item{mE}{an integer. Used when fitting the factor analytical model if covModel = 3.
See \code{\link{EMFA}}.}

\item{CmHet}{a boolean. Used when fitting the factor analytical model if covModel = 3.
See \code{\link{EMFA}}.}

\item{DmHet}{a boolean. Used when fitting the factor analytical model if covModel = 3.
See \code{\link{EMFA}}.}

\item{stopIfDecreasing}{a boolean. Used when fitting the factor analytical model if covModel = 3.
See \code{\link{EMFA}}.}

\item{computeLogLik}{a boolean. Used when fitting the factor analytical model if covModel = 3.
See \code{\link{EMFA}}.}

\item{Vg}{an optional matrix with genotypic variance components. Vg should have row names
column names corresponding to the column names of Y. It may contain additional rows and colums
which will be ignored. Ignored if fitVarComp = \code{TRUE}.}

\item{Ve}{an optional matrix with environmental variance components. Ve should have row names
column names corresponding to the column names of Y. It may contain additional rows and colums
which will be ignored. Ignored if fitVarComp = \code{TRUE}.}

\item{reduceK}{if \code{TRUE} the kinship matrix is reduced. See \code{\link{reduceKinship}}}

\item{nPca}{an integer giving the number of Pcas used whe reducing the kinship matrix.
Ignored if reduceK = \code{FALSE}}
}
\value{
a list containing the following items:
\itemize{
\item{\code{Vg} a matrix with genotypic variance compontents.}
\item{\code{Ve} a matrix with environmental variance compontents.}
\item{\code{M} a matrix of effect size estimates.}
\item{\code{TStat} a matrix of t-statistics.}
\item{\code{results} a vector of p-values.}
\item{\code{resultsWald} a vector of p-values for the wald test.}
\item{\code{MExtended} a matrix of snp information, lod scores, lod scores for the Wald
test and effect size estimates.}
\item{\code{TStatExtended} a matrix of snp information, lod scores, lod scores for the Wald
test and t-statistics.}
}
}
\description{
run GWAS
}
\references{
Dahl et al. (2014). Network inference in matrix-variate Gaussian models with
non-indenpent noise.

Zhou, X. and Stephens, M. (2014). Efficient multivariate linear mixed model algorithms for
genome-wide association studies.
}
