% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSingleTraitGwas.R
\name{runSingleTraitGwas}
\alias{runSingleTraitGwas}
\title{run single trait GWAS}
\usage{
runSingleTraitGwas(gData, GLSMethod = 3, sizeIncludedRegion = 0, minR2,
  useMAF = TRUE, MAF = 0.05, MAC = 10)
}
\arguments{
\item{gData}{an object of class \code{gData}}

\item{GLSMethod}{an integer indicating the software used to estimate the marker effects.
\enumerate{
\item{scan_GLS (not available when there are heterozygotes) - NOT IMPLEMENTED YET}
\item{Fast-LMM - NOT IMPLEMENTED YET}
\item{within R}
\item{within R, with chromosome specific kinship matrices; similar to the approach of Rincent et al.
(genetics, 2014)}
}}

\item{sizeIncludedRegion}{an integer. Should the results for SNPs close to significant SNPs
be included? If so, the size of the region in centimorgan or base pairs. Otherwise 0.}

\item{minR2}{A numerical value between 0 and 1. Restricts the SNPs included in the region close
to significant SNPs to only those SNPs that are in sufficient LD with the significant snp,
where LD is measured in terms of r^2. If for example \code{sizeIncludedRegion} = 200000
and \code{minR2} = 0.5, then for every significant SNP also those SNPs whose LD (r^2) with the
significant SNP is at least 0.5 AND which are at most 200kb away from this significant snp are
included. Ignored if \code{sizeIncludedRegion} = 0.}

\item{useMAF}{should the minor allele frequency be used for selecting SNPs for the analysis.
If \code{FALSE} the minor allele count is used.}

\item{MAF}{A numerical value between 0 and 1. SNPs with minor allele frequency below
this value are not taken into account for the analysis, i.e. p-values and effect sizes are
put to missing (NA). Ignored if \code{useMAF} is \code{FALSE}.}

\item{MAC}{A numerical value. SNPs with minor allele count below this value are not
taken into account for the analysis, i.e. p-values and effect sizes are put to missing (NA).
Ignored if \code{useMAF} is \code{TRUE}.}
}
\description{
run single trait GWAS
}
\references{
Rincent et al. (2014) ....
}
