% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runSingleTraitGwasIBD.R
\name{runSingleTraitGwasIBD}
\alias{runSingleTraitGwasIBD}
\title{Perform IBD based QTL mapping for single trait.}
\usage{
runSingleTraitGwasIBD(gData, traits = NULL, environments = NULL,
  covar = NULL, snpCov = NULL, kin = NULL,
  kinshipMethod = c("multiAllKin"), remlAlgo = c("EMMA", "NR"),
  GLSMethod = c("single", "multi"), MAF = 0.01,
  genomicControl = FALSE, thrType = c("bonf", "fixed", "small"),
  alpha = 0.05, LODThr = 4, nSnpLOD = 10, sizeInclRegion = 0,
  minR2 = 0.5, ref = 1, nCores = NULL)
}
\arguments{
\item{gData}{An object of class \code{gData} containing at least \code{map},
\code{markers} and \code{pheno}.}

\item{traits}{A vector of traits on which to run GWAS. These can be either
numeric indices or character names of columns in \code{pheno}. If \code{NULL},
GWAS is run on all traits.}

\item{environments}{A vector of environments on which to run GWAS. These can
be either numeric indices or character names of list items in \code{pheno}.
If \code{NULL}, GWAS is run for all environments. GWAS is run for the
selected environments in sequential order.}

\item{covar}{An optional vector of covariates taken into account when
running GWAS. These can be either numeric indices or character names of
columns in \code{covar} in \code{gData}. If \code{NULL} no covariates are
used.}

\item{snpCov}{An optional character vector of snps to be included as
covariates.}

\item{kin}{An optional kinship matrix or list of kinship matrices. These
matrices can be from the \code{matrix} class as defined in the base package
or from the \code{dsyMatrix} class, the class of symmetric matrices in the
Matrix package.\cr
If \code{GLSMethod} = "single" then one matrix should be provided, if
\code{GLSMethod} = "multi", a list of chromosome specific matrices of length
equal to the number of chromosomes in \code{map} in \code{gData}.\cr
If \code{NULL} then matrix \code{kinship} in \code{gData} is used. \cr
If both \code{kin} is provided and \code{gData} contains a matrix
\code{kinship} then \code{kin} is used.}

\item{kinshipMethod}{An optional character indicating the method used for
calculating the kinship matrix(ces). Currently "astle" (Astle and Balding,
2009), "IBS" and "vanRaden" (VanRaden, 2008) are supported. If a
kinship matrix is supplied either in \code{gData} or in parameter \code{kin},
\code{kinshipMethod} is ignored.}

\item{remlAlgo}{A character string indicating the algorithm used to estimate
the variance components. Either \code{EMMA}, for the EMMA algorithm, or
\code{NR}, for the Newton-Raphson algorithm.}

\item{GLSMethod}{A character string indicating the method used to estimate
the marker effects. Either \code{single} for using a single kinship matrix,
or \code{multi} for using chromosome specific kinship matrices.}

\item{MAF}{The minor allele frequency (MAF) threshold used in GWAS. A
numerical value between 0 and 1. SNPs with MAF below this value are not taken
into account in the analysis, i.e. p-values and effect sizes are put to
missing (\code{NA}). Ignored if \code{useMAF} is \code{FALSE}.}

\item{genomicControl}{Should genomic control correction as in Devlin and
Roeder (1999) be applied?}

\item{thrType}{A character string indicating the type of threshold used for
the selection of candidate loci. Either \code{bonf} for using the
Bonferroni threshold, a LOD-threshold of \eqn{-log10(alpha/p)}, where p is
the number of markers and alpha can be specified in \code{alpha},
\code{fixed} for a self-chosen fixed LOD-threshold, specified in \code{LODThr}
or \code{small}, the LOD-threshold is chosen such as the SNPs with the
\code{nSnpLOD} smallest p-values are selected. \code{nSnpLOD} can be
specified.}

\item{alpha}{A numerical value used for calculating the LOD-threshold for
\code{thrType} = "bonf".}

\item{LODThr}{A numerical value used as a LOD-threshold when
\code{thrType} = "fixed".}

\item{nSnpLOD}{A numerical value indicating the number of SNPs with the
smallest p-values that are selected when \code{thrType} = "small".}

\item{sizeInclRegion}{An integer. Should the results for SNPs close to
significant SNPs be included? If so, the size of the region in centimorgan
or base pairs. Otherwise 0.}

\item{minR2}{A numerical value between 0 and 1. Restricts the SNPs included
in the region close to significant SNPs to only those SNPs that are in
sufficient Linkage Disequilibruim (LD) with the significant snp, where LD
is measured in terms of \eqn{R^2}. If for example \code{sizeInclRegion} = 200000
and \code{minR2} = 0.5, then for every significant SNP also those SNPs whose
LD (\eqn{R^2}) with the significant SNP is at least 0.5 AND which are at
most 200kb away from this significant snp are included. Ignored if
\code{sizeInclRegion} = 0.}

\item{ref}{A numerical value indicating the founder allele to be used as
reference allele. Effects for the other founder allele are computed based on
this reference allele.}

\item{nCores}{A numerical value indicating the number of cores to be used by
the parallel part of the algorithm. If \code{NULL} the number of cores used
will be equal to the number of cores available on the machine - 1.}
}
\value{
An object of class \code{\link{GWAS}}.
}
\description{
\code{runSingleTraitGwasIBD} QTL mapping based on IBD probabilities.
}
\seealso{
\code{\link{GWAS}}, \code{\link{kinship}}
}
