% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMFA.R
\name{startValPW}
\alias{startValPW}
\title{Helper function for computing starting values}
\usage{
startValPW(C, m, p)
}
\arguments{
\item{C}{A precision matrix.}

\item{m}{An integer, the order of the model.}

\item{p}{An integer, the number of traits or genotypes.}
}
\description{
Helper function for computing starting values for P and W.
}
\keyword{internal}
