% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statgenPipeline.R
\docType{package}
\name{statgenPipeline-package}
\alias{statgenPipeline}
\alias{statgenPipeline-package}
\title{statgenPipeline: Genome Wide Association Studies in R}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Bart-Jan van Rossum \email{bart-jan.vanrossum@wur.nl}

Other contributors:
\itemize{
  \item Willem Kruijer \email{willem.kruijer@wur.nl} [contributor]
  \item Maikel Verouden \email{maikel.verouden@wur.nl} [contributor]
}

}
\keyword{internal}
