% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.GWAS.R
\name{summary.GWAS}
\alias{summary.GWAS}
\title{Summary function for the class \code{GWAS}}
\usage{
\method{summary}{GWAS}(object, ..., environments = NULL)
}
\arguments{
\item{object}{object of class \code{GWAS}}

\item{...}{not used}

\item{environments}{a vector of strings or numeric indices indicating for which environment the
summary should be made. If \code{NULL} a summary is made for all environments.}
}
\description{
Gives a summary for an object of S3 class \code{GWAS}.
}
