% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracePInvDiag.R
\name{tracePInvDiag}
\alias{tracePInvDiag}
\title{tracePInvDiag}
\usage{
tracePInvDiag(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a matrix containing the column sums of \eqn{(I + x \otimes y)^{-1}}
}
\description{
Helper function for quickly computing column sums of \eqn{(I + x \otimes y)^{-1}}. Used in the
penalized EM algorith.
}
\examples{
tracePInvDiag(1:2, 1:3)
}
