% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updateFAHomVarCPP}
\alias{updateFAHomVarCPP}
\title{Update W and P in EMFA algorithm for homogeneous variance.}
\usage{
updateFAHomVarCPP(s, wNew, pNew, m, maxDiag = 10000)
}
\arguments{
\item{s}{A p x p sample covariance matrix.}

\item{m}{An integer. The order of the model.}

\item{maxDiag}{A numerical value for the maximum value of sigma2.}
}
\description{
Update W and P used in the iteration process in the EMFA algorithm in case
the variance is homogeneous.
}
\keyword{internal}
