% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{updatePrec}
\alias{updatePrec}
\title{Helper function for updating precision matrix.}
\usage{
updatePrec(m, nc, omega, w, p, wNew, pNew, cNew, het, maxDiag)
}
\arguments{
\item{m}{An integer, the order of the model.}

\item{nc}{An integer, the number of traits or genotypes.}

\item{omega}{A computed matrix for the current step in the algoritm.}

\item{w}{A model matrix for the current step in the algorithm.}

\item{p}{A model matrix for the current step in the algorithm.}

\item{wNew}{A pointer to the updated model matrix for w.}

\item{pNew}{A pointer to the updated model matrix for p.}

\item{cNew}{A pointer to the updated matrix c.}

\item{het}{Should an extra diagonal part be added in the model for the
precision matrix.}

\item{maxDiag}{A numerical value for the maximum value of sigma2.}
}
\description{
Helper function for updating the precision matrices in the EMFA algorithm.
}
\keyword{internal}
