% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecInvDiag.R
\name{vecInvDiag}
\alias{vecInvDiag}
\title{vecInvDiag}
\usage{
vecInvDiag(x, y)
}
\arguments{
\item{x}{a numeric vector}

\item{y}{a numeric vector}
}
\value{
a matrix defined by \eqn{(I + x \otimes y)^{-1}}
}
\description{
Helper function for quickly computing \eqn{(I + x \otimes y)^{-1}}. Used in the penalized EM algorith.
}
\examples{
vecInvDiag(1:2, 1:3)

}
