% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMFA.R
\name{EMFA}
\alias{EMFA}
\title{Factor analytic variation of EM algoritm}
\usage{
EMFA(Y, K, X = matrix(rep(1, nrow(K))), CmHet = FALSE, DmHet = FALSE,
  tolerance = 1e-04, maxIter = 300L, CmStart = NULL, DmStart = NULL,
  mG = 1, mE = 1, maxDiag = 10000, prediction = TRUE,
  stopIfDecreasing = FALSE, computeLogLik = FALSE)
}
\arguments{
\item{Y}{an n x p matrix of observed phenotypes, on p traits or environments for n individuals.
No missing values are allowed.}

\item{K}{an n x n kinship matrix.}

\item{X}{an n x c covariate matrix, c being the number of covariates and n being the number
of genotypes. c has to be at least one (typically an intercept). No missing values are allowed.
If not provided a vector of 1s is used.}

\item{CmHet}{if \code{TRUE} an extra diagonal part is added in the model for the
precision matrix Cm.}

\item{DmHet}{if \code{TRUE} an extra diagonal part is added in the model for the
precision matrix Dm.}

\item{tolerance}{a numerical value. The iterating process stops if the difference in log-likelihood
between two iterations gets smaller than the tolerance.}

\item{maxIter}{a numerical value for the maximum number of iterations.}

\item{CmStart}{a p x p matrix containing starting values for the precision matrix Cm.}

\item{DmStart}{a p x p matrix containing starting values for the precision matrix Dm.}

\item{mG}{an integer. The order of the genetic part of the model.}

\item{mE}{an integer. The order of the environmental part of the model.}

\item{maxDiag}{a numical value. The maximal value of the diagonal elements in the precision matrices
Cm and Dm (ignoring the low-rank part W W^t)}

\item{prediction}{if \code{TRUE} values for Y are predicted and returned.}

\item{stopIfDecreasing}{if \code{TRUE} the iterating process stops if after 50 iterations the
log-likelihood decreases between two iterations.}

\item{computeLogLik}{if \code{TRUE} log-likelihood is returned.}
}
\value{
A list containing results of the algoritm.
}
\description{
Implementation of the factor analytic variation of the EM algoritm as proposed by Dahl et al.
}
\references{
Dahl et al. (2014). Network inference in matrix-variate Gaussian models with
non-indenpent noise.
}
