% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateEffects.R
\name{estimateEffects}
\alias{estimateEffects}
\title{Estimates for covariates.}
\usage{
estimateEffects(X, x = NULL, Y, VInvArray, returnAllEffects = TRUE)
}
\arguments{
\item{X}{is a c x n covariate matrix, c being the number of covariates and n being the number
of genotypes. c has to be at least one (typically an intercept). See details.}

\item{x}{is an optional additional covariate, a numeric vector of length n. See details.}

\item{Y}{is a p x n matrix of observed phenotypes, on p traits or environments for n genotypes.
See details.}

\item{VInvArray}{an n x p x p dimensional array obtained as an output from the function
\code{\link{makeVInvArray}}. It contains for each genotype i the p x p matrix \eqn{v_l ^ {-1}} (in
the notation of Zhou and Stephens)}

\item{returnAllEffects}{If \code{FALSE} only the p effect estimates for the additional covariate
are returned, (the first p * c are actually computed, but not returned). If \code{TRUE},
also the effect estimates and standard errors for the other covariates (i.e. those in X) are returned.
If no x is provided, \code{returnAllEffects} is automatically set to \code{TRUE}.}
}
\value{
A list containing the estimates, the standard errors of the estimates and if an value for
x is input
}
\description{
Compute the estimates and standard errors for the covariates in the input matrix X. Optionally an
additional covariate x can be estimated.
}
\details{
No missing values are allowed in X, Y and x.\cr
It is assumed that X, Y and x have already been rotated by Uk, where Uk is such that
the kinship matrix K equals \eqn{K = Uk * Dk * t(Uk)}.\cr
The original X, Y and x are right multiplied by Uk, e.g. \code{Y <- Y * Uk}. See Zhou
and Stephens 2014, supplement.\cr
It is these rotated versions that are the input of this function.
}
\references{
Zhou, X. and Stephens, M. (2014). Efficient mutivariate linear mixed model algorithms for
genome-wide association studies.
}
