% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomicPrediction.R
\name{genomicPrediction}
\alias{genomicPrediction}
\title{Perform genomic prediction}
\usage{
genomicPrediction(
  gData,
  traits = NULL,
  environments = NULL,
  training,
  useMAF = TRUE,
  MAF = 0.01,
  MAC = 10,
  keep = NULL
)
}
\arguments{
\item{gData}{An object of class \code{gData} containing at least \code{map},
\code{markers} and \code{pheno}.}

\item{traits}{A vector of traits on which to run GWAS. These can be either
numeric indices or character names of columns in \code{pheno}. If \code{NULL},
GWAS is run on all traits.}

\item{environments}{A vector of environments on which to run GWAS. These can
be either numeric indices or character names of list items in \code{pheno}.
If \code{NULL}, GWAS is run for all environments. GWAS is run for the
selected environments in sequential order.}

\item{training}{A character vector with names of genotypes to use as
training data for the prediction.}

\item{useMAF}{Should the minor allele frequency be used for selecting SNPs
for the analysis. If \code{FALSE}, the minor allele count is used instead.}

\item{MAF}{The minor allele frequency (MAF) threshold used in GWAS. A
numerical value between 0 and 1. SNPs with MAF below this value are not taken
into account in the analysis, i.e. p-values and effect sizes are put to
missing (\code{NA}). Ignored if \code{useMAF} is \code{FALSE}.}

\item{MAC}{A numerical value. SNPs with minor allele count below this value
are not taken into account for the analysis, i.e. p-values and effect sizes
are set to missing (\code{NA}). Ignored if \code{useMAF} is \code{TRUE}.}

\item{keep}{A character vector with names of columns in \code{covar} in
\code{gData} that should be included in the output.}
}
\value{
A data.frame containing predictions for genotypes per trait.
Columns in \code{keep} are merged from covar in gData and included in
the output.
}
\description{
Perform genomic prediction
}
\references{
Piepho, H.-P., Möhring, J., Schulz-Streeck, T. and Ogutu, J. O.
(2012), A stage-wise approach for the analysis of multi-environment trials.
Biom. J., 54: 844–860. doi:10.1002/bimj.201100219
}
