% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVArray.R
\name{makeVArray}
\alias{makeVArray}
\alias{makeVInvArray}
\title{Create array of (inverted) variance matrices}
\usage{
makeVArray(Vg, Ve, Dk)

makeVInvArray(Vg, Ve, Dk)
}
\arguments{
\item{Vg}{is a p x p symmetric matrix of genetic variance components}

\item{Ve}{is a p x p symmetric matrix of environmental variance components}

\item{Dk}{is a diagonal n x n matrix containing the eigenvalues obtained by the eigendecomposition
of the kinship matrix K: \eqn{K = Uk * Dk * t(Uk)}}
}
\value{
An array of n p x p matrices \eqn{v_l} where \eqn{v_l = Dk_{l,l} * Vg + Ve \forall
l = 1, ..., n}.\cr
When using \code{makeVInvArray} the output matrices are inverted.
}
\description{
As in Zhou and Stephens, supplement page 13, create the array of (inverted) variance matrices for the
transformed genotypes per individual as given by equation 4.
}
\references{
Zhou, X. and Stephens, M. (2014). Efficient mutivariate linear mixed model algorithms for
genome-wide association studies, supplement 13-14.
}
