% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manhattanPlot.R
\name{manhattanPlot}
\alias{manhattanPlot}
\title{Create a manhattan plot.}
\usage{
manhattanPlot(xValues, yValues, map, fileName = "", jpegPlot = TRUE,
  xLab = "Chromosomes", yLab = expression(-log[10](p)), title = "",
  plotType = "l", xSig = integer(), xEffects = integer(),
  colPalette = rep(c("royalblue", "maroon"),
  50)[1:length(levels(factor(map$chromosome)))], chrBoundaries = 0,
  yThr = NULL, signPointsThickness = 0.6)
}
\arguments{
\item{xValues}{a vector of cumulative marker positions.}

\item{yValues}{a vector of LOD-scores.}

\item{map}{a dataframe with at least the columns chromosome, the number of the chromosome
and cum.position, the cumulative position of the snp the cumulative position of the snp
starting from the first chromosome.}

\item{fileName}{name of the outputfile that is created. If left empty the plot is written
to the screen.}

\item{jpegPlot}{should a jpeg file be produced? If \code{FALSE} a pdf file is produced.}

\item{xLab}{x-axis label.}

\item{yLab}{y-axis label.}

\item{title}{title of the plot}

\item{plotType}{lines ("l") or dots ("d" or "p")}

\item{xSig}{vector of integers, indicating which components in the vectors xValues and
yValues are significant.}

\item{xEffects}{vector of integers, indicating which components in the vector xValues
correspond to a real (known) effect.}

\item{colPalette}{color palette used for plotting.}

\item{chrBoundaries}{vector of chromosome boundaries, i.e. x-values on the same scale as xValues.}

\item{yThr}{LOD-threshold.}

\item{signPointsThickness}{thickness of the points that are false/true positives/negatives.}
}
\value{
a LOD-profile with LOD-scores per snip. Markers declared significant get a red dot,
markers with a real effect get a blue dot. If both significant and real effects are given
false positives get an orange dot, true negatives a yellow dot and true positives a green dot.
}
\description{
Given vectors of marker positions and corresponding LOD-scores plot a LOD-profile.
Significant markers can be highlighted with red dots. If there are previously known
marker effect also false positives and true negatives will be marked.
}
