% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqPlot.R
\name{qqPlot}
\alias{qqPlot}
\title{qq-plot of observed versus expected LOD-scores}
\usage{
qqPlot(pValues, title = "QQ-plot", fileName = "", jpegPlot = TRUE)
}
\arguments{
\item{pValues}{a numeric vector of pValues. Missings are ignored when plotting.}

\item{title}{main title for the plot}
}
\description{
Given a vector of pvalues, generate a qq-plot of observed LOD-scores versus expected LOD-scores.
Code taken from Segura et al. and adapted
}
\references{
Segura et al. (2012) An efficient multi-locus mixed model approach for genome-wide
association studies in structured populations
}
