% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceKinship.R
\name{reduceKinship}
\alias{reduceKinship}
\title{Reduce the kinship matrix}
\usage{
reduceKinship(K, nPca)
}
\arguments{
\item{nPca}{an integer, the number of eigenvectors used for reducing the k
inship matrix.}
}
\value{
the reduced kinship matrix
}
\description{
The kinship matrix is reduced using nPca eigenvectors of K.
}
\keyword{internal}
