% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMultiTraitGwas.R
\name{runMultiTraitGwas}
\alias{runMultiTraitGwas}
\title{Perform multi-trait GWAS}
\usage{
runMultiTraitGwas(gData, environments = NULL, covar = NULL,
  snpCov = NULL, kin = NULL, kinshipMethod = c("astle", "GRM", "IBS",
  "vanRaden"), GLSMethod = c("single", "multi"), subsetMarkers = FALSE,
  markerSubset = "", MAF = 0.01, fitVarComp = TRUE,
  covModel = c("unst", "pw", "fa"), VeDiag = TRUE, tolerance = 1e-06,
  maxIter = 2e+05, maxDiag = 10000, mG = 1, mE = 1, CmHet = TRUE,
  DmHet = TRUE, stopIfDecreasing = TRUE, computeLogLik = TRUE,
  Vg = NULL, Ve = NULL, reduceK = FALSE, nPca = NULL,
  estCom = FALSE, parallel = FALSE, nCores = NULL)
}
\arguments{
\item{gData}{An object of class \code{gData} containing at least \code{map},
\code{markers} and \code{pheno}.}

\item{environments}{A vector of environments on which to run GWAS. These can
be either numeric indices or character names of list items in \code{pheno}.
If \code{NULL} GWAS is run for all environments.}

\item{covar}{An optional vector of covariates taken into account when
running GWAS. These can be either numeric indices or character names of
columns in \code{covar} in \code{gData}. If \code{NULL} no covariates are
used.}

\item{snpCov}{An optional character vector of snps to be included as
covariates.}

\item{kin}{An optional kinship matrix or list of kinship matrices. These
matrices can be from the \code{matrix} class as defined in the base package
or from the \code{dsyMatrix} class, the class of symmetric matrices in the
Matrix package.\cr
If \code{GLSMethod} = "single" then one matrix should be provided, if
\code{GLSMethod} = "multi" a list of chromosome specific matrices of lenght
equal to the number of chromosomes in \code{map} in \code{gData}.\cr
If \code{NULL} then matrix \code{kinship} in \code{gData} is used. \cr
If both \code{kin} is provided and \code{gData} contains a matrix
\code{kinship} then \code{kin} is used.}

\item{kinshipMethod}{An optional character indicating the method used for
calculating the kinship matrix(ces). Currently "astle" (Astle and Balding,
2009), "GRM", "IBS" and "vanRaden" (VanRaden, 2008) are supported. If a
kinship matrix is supplied either in \code{gData} or in parameter \code{kin}
\code{kinshipMethod} is ignored.}

\item{GLSMethod}{A character string indicating the method used to estimate
the marker effects. Either \code{single} for using a single kinship matrix.
or \code{multi} for using chromosome specific kinship matrices.}

\item{subsetMarkers}{Should the marker data be subsetted?}

\item{markerSubset}{A numeric or character vector used for subsetting the
markers. Ignored if subsetMarkers = \code{FALSE}.}

\item{MAF}{A numerical value between 0 and 1. SNPs with minor allele
frequency below this value are not taken into account for the analysis,
i.e. p-values and effect sizes are put to missing (NA). Ignored if
\code{useMAF} is \code{FALSE}.}

\item{fitVarComp}{Should the variance components be fitted? If \code{FALSE}
they should be supplied in Vg and Ve}

\item{covModel}{An character string indictating the model used when fitting
the variance components. Either \code{unst} for unstructured for both Vg and
Ve (as in Zhou and Stephens (2014)), \code{pw} for unstructered for both Vg
and Ve (pairwise, as in Furlotte and Eskin (2013)) or \code{fa} for
factor-analytic for both Vg and Ve.\cr
Ignored if \code{fitVarComp} = \code{FALSE}}

\item{VeDiag}{Should there be environmental correlations if covModel = "unst"
or "pw"? If traits are measured on the same individuals put \code{FALSE}.}

\item{tolerance}{A numerical value. Used when fitting the factor analytical
model if covModel = "fa". See \code{\link{EMFA}}.}

\item{maxIter}{An integer. Used when fitting the factor analytical model if
covModel = "fa". See \code{\link{EMFA}}.}

\item{maxDiag}{A numerical value. Used when fitting the factor analytical
model if covModel = "fa". See \code{\link{EMFA}}.}

\item{mG}{An integer. Used when fitting the factor analytical model if
covModel = "fa". See \code{\link{EMFA}}.}

\item{mE}{An integer. Used when fitting the factor analytical model if
covModel = "fa". See \code{\link{EMFA}}.}

\item{CmHet}{A boolean. Used when fitting the factor analytical model if
covModel = "fa". See \code{\link{EMFA}}.}

\item{DmHet}{A boolean. Used when fitting the factor analytical model if
covModel = "fa". See \code{\link{EMFA}}.}

\item{stopIfDecreasing}{A boolean. Used when fitting the factor analytical
model if covModel = "fa". See \code{\link{EMFA}}.}

\item{computeLogLik}{A boolean. Used when fitting the factor analytical model
if covModel = "fa". See \code{\link{EMFA}}.}

\item{Vg}{An optional matrix with genotypic variance components. Vg should
have row names column names corresponding to the column names of Y. It may
contain additional rows and colums which will be ignored. Ignored if
fitVarComp = \code{TRUE}.}

\item{Ve}{An optional matrix with environmental variance components. Ve
should have row names column names corresponding to the column names of Y.
It may contain additional rows and colums which will be ignored. Ignored if
fitVarComp = \code{TRUE}.}

\item{reduceK}{Should the kinship matrix be reduced? See
\code{\link{reduceKinship}}}

\item{nPca}{An integer giving the number of Pcas used whe reducing the
kinship matrix. Ignored if reduceK = \code{FALSE}.}

\item{estCom}{Should the common SNP-effect model be fitted?}

\item{parallel}{Should the computation of variance components be done in
parallel? Only used if \code{covModel = "pw"}. A parallel computing
environment has to be setup by the user.}

\item{nCores}{A numerical value indicating the number of cores to be used by
the parallel part of the algorithm. If \code{NULL} the number of cores used
will be equal to the number of cores available on the machine - 1.}
}
\value{
An object of class \code{\link{GWAS}}.
}
\description{
\code{runMultiTraitGwas} performs a multi-trait Genome Wide Association
Study (GWAS) on phenotypic and genotypic data contained in a \code{gData}
object.
}
\references{
Dahl et al. (2013). Network inference in matrix-variate Gaussian
models with non-independent noise. arXiv preprint arXiv:1312.1622.

Furlotte, N.A. and Eskin, E. (2015). Efficient multiple-trait
association and estimation of genetic correlation using the matrix-variate
linear mixed model. Genetics, May 2015, Vol.200-1, p. 59-68.

Kruijer et al. (2015) Marker-based estimation of heritability in
immortal populations. Genetics. February 2015, Vol. 199-2, p. 379-398.

Zhou, X. and Stephens, M. (2014). Efficient multivariate linear
mixed model algorithms for genome-wide association studies. Nature Methods,
February 2014, Vol. 11, p. 407–409.
}
