% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFA.R
\name{updateFA}
\alias{updateFA}
\title{updateFA}
\usage{
updateFA(Y, WStart = NULL, m = ifelse(is.null(WStart), 2, ncol(WStart)),
  PStart = NULL, hetVar = FALSE, maxDiag = 10000, tolerance = 1e-04,
  maxIter = 100L, printProgress = FALSE)
}
\arguments{
\item{Y}{a n x p matrix or dataframe.}

\item{WStart}{a p x p matrix or dataframe containing starting values for W.}

\item{m}{an integer. The order of the model.}

\item{PStart}{a p x p matrix or dataframe containing starting values for P.}

\item{hetVar}{should an extra diagonal part be added in the model for the precision matrix?}

\item{maxDiag}{a numerical value for the maximum value of the diagonal of P.}

\item{tolerance}{a numerical value. The iterating process stops if the sum of the difference for P
and W between two steps gets lower than this value.}

\item{maxIter}{a numerical value for the maximum number of iterations.}

\item{printProgress}{should progress be printed during iterations?}
}
\value{
a list containing the new matrices W and P after the iteration process and the number
of iterations.
}
\description{
updateFA
}
