% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMMA.R
\name{updateFA}
\alias{updateFA}
\title{Update W and P in EMFA algorithm}
\usage{
updateFA(Y, WStart = NULL, m = ifelse(is.null(WStart), 2,
  ncol(WStart)), PStart = NULL, hetVar = FALSE, maxDiag = 10000,
  tolerance = 1e-04, maxIter = 100L, printProgress = FALSE)
}
\arguments{
\item{Y}{An n x p matrix or data.frame.}

\item{WStart}{A p x p matrix or data.frame containing starting values for W.}

\item{m}{An integer. The order of the model.}

\item{PStart}{A p x p matrix or data.frame containing starting values for P.}

\item{hetVar}{Should an extra diagonal part be added in the model for the
precision matrix?}

\item{maxDiag}{A numerical value for the maximum value of the diagonal of P.}

\item{tolerance}{A numerical value. The iterating process stops if the sum
of the difference for P and W between two steps gets lower than this value.}

\item{maxIter}{A numerical value for the maximum number of iterations.}

\item{printProgress}{Should progress be printed during iterations?}
}
\value{
A list containing the new matrices W and P after the iteration
process and the number of iterations.
}
\description{
Updata W and P used in the iteration process in the EMFA algorithm.
}
\keyword{internal}
