% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateFAHomVar.R
\name{updateFAHomVar}
\alias{updateFAHomVar}
\title{Update W and P in EMFA algorithm for homogeneous variance.}
\usage{
updateFAHomVar(Y = NULL, S = NULL, m, maxDiag = 10000)
}
\arguments{
\item{Y}{An n x p matrix of observed phenotypes, on p traits or environments
for n individuals. No missing values are allowed.}

\item{S}{An p x p sample covariance matrix.}

\item{m}{An integer. The order of the model.}

\item{maxDiag}{A numerical value for the maximum value of sigma2.}
}
\value{
A list containing the new value for the matrices W and P.
}
\description{
Updata W and P used in the iteration process in the EMFA algorithm in case
the variance is homogeneous.
}
\keyword{internal}
