% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EMFA.R
\name{updatePrec}
\alias{updatePrec}
\title{Helper function for updating precision matrix.}
\usage{
updatePrec(m, p, Omega, W, P, het, maxDiag)
}
\arguments{
\item{m}{An integer, the order of the model.}

\item{p}{An integer, the number of traits or genotypes.}

\item{Omega}{A computed matrix for the current step in the algoritm.}

\item{W}{A model matrix for the current step in the algorithm.}

\item{P}{A model matrix for the current step in the algorithm.}

\item{het}{Should an extra diagonal part be added in the model for the
precision matrix.}

\item{maxDiag}{A numerical value for the maximum value of sigma2.}
}
\description{
Helper function for updating the precision matrices in the EMFA algorithm.
}
\keyword{internal}
